drm/amdgpu: Make noretry the default on yellow carp
Performing kfd page fault tests on the yellow carp platform will fail.

From the scan data after the failure, it can be found that a
nack=0x1(retry fault) is returned. but we did not enable the interrupts
for retry faults in the code.

So we need to set noretry = 1 like the above ASICs.

Signed-off-by: chen gong <curry.gong@amd.com>
Reviewed-by: Aaron Liu <aaron.liu@amd.com>
Acked-by: Huang Rui <ray.huang@amd.com>