drm/amd/display: Introduce KUnit tests for fixed31_32 library
The fixed31_32 library performs a lot of the mathematical operations
involving fixed-point arithmetic and the conversion of integers to
fixed-point representation.

This unit tests intend to assure the proper functioning of the basic
mathematical operations of fixed-point arithmetic, such as
multiplication, conversion from fractional to fixed-point number,
and more.

Tested with kunit-tools and IGT (with some patches[1]).

1) Kunit tools:

$ ./tools/testing/kunit/kunit.py run \
	--arch=x86_64 \
	--kunitconfig=drivers/gpu/drm/amd/display/tests/

2) IGT (aided by virtme and KW)

$ virtme/virtme-configkernel --arch=x86_64 --defconfig
$ linux/scripts/config -e KUNIT -m DRM_AMDGPU -e DRM_AMD_DC \
		       -m AMD_DAL_BASICS_KUNIT_TEST
$ kw build
$ virtme/virtme-run --mods=auto --kdir linux

Now, inside QEMU:

$ $PATH_TO_IGT/igt-gpu-tools/scripts/run-tests.sh -t ".*kunit.*" -r
/tmp

Signed-off-by: Tales Aparecida <tales.aparecida@gmail.com>

Link: https://gitlab.freedesktop.org/tales-aparecida/igt-gpu-tools/-/tree/kunit-prototype-v5.19-fixed31_32 [1]